requirements.txt
# local package
-e .

# external requirements
flake8
python-dotenv>=0.5.1

pandas
numpy
scikit-learn
pyyaml

#torch
#transformers
#atasets

#mlflow
great_expectations
#dvc

#fastapi
#uvicorn

python-dotenv
pyyaml

pytest

